+++ src/beforeBuild.ts
@@ -15,14 +15,7 @@
  */
 
 import assert from "node:assert";
-import {
-  access,
-  cp,
-  mkdir,
-  readFile,
-  unlink,
-  writeFile,
-} from "node:fs/promises";
+import { access, cp, mkdir, readFile, rm, writeFile } from "node:fs/promises";
 import { join, sep } from "node:path";
 
 import { compile } from "./lib/aqua.js";
@@ -139,13 +132,9 @@ const compileInstallationSpellAqua = async (tracing = false) => {
   );
 };
 
-(async () => {
-  try {
-    await unlink(VERSIONS_DIR_PATH);
-  } catch {}
-
+void (async () => {
+  await rm(VERSIONS_DIR_PATH, { recursive: true, force: true });
   await mkdir(VERSIONS_DIR_PATH, { recursive: true });
-
   await cp("package.json", join(VERSIONS_DIR_PATH, "cli.package.json"));
 
   await cp(
@@ -153,10 +142,7 @@ const compileInstallationSpellAqua = async (tracing = false) => {
     join(VERSIONS_DIR_PATH, "js-client.package.json")
   );
 
-  try {
-    await unlink(COMPILED_AQUA_PATH);
-  } catch {}
-
+  await rm(COMPILED_AQUA_PATH, { recursive: true, force: true });
   await mkdir(COMPILED_INSTALLATION_SPELL_AQUA_PATH, { recursive: true });
 
   await mkdir(COMPILED_INSTALLATION_SPELL_AQUA_WITH_TRACING_PATH, {
@@ -165,6 +151,4 @@ const compileInstallationSpellAqua = async (tracing = false) => {
 
   await compileInstallationSpellAqua();
   await compileInstallationSpellAqua(true);
-})().catch((e) => {
-  return console.error(e);
-});
+})();