+++ src/commands/dependency/versions.ts
@@ -14,12 +14,16 @@
  * limitations under the License.
  */
 
+import { Flags } from "@oclif/core";
 import { yamlDiffPatch } from "yaml-diff-patch";
 
 import { BaseCommand, baseFlags } from "../../baseCommand.js";
 import { commandObj } from "../../lib/commandObj.js";
 import { CLI_NAME } from "../../lib/const.js";
-import { resolveDependencies } from "../../lib/helpers/package.js";
+import {
+  getRecommendedDependencies,
+  resolveDependencies,
+} from "../../lib/helpers/package.js";
 import { initCli } from "../../lib/lifeCycle.js";
 import CLIPackageJSON from "../../versions/cli.package.json" assert { type: "json" };
 import JSClientPackageJSON from "../../versions/js-client.package.json" assert { type: "json" };
@@ -27,17 +31,36 @@ import versions from "../../versions.json" assert { type: "json" };
 
 export default class Versions extends BaseCommand<typeof Versions> {
   static override aliases = ["dependency:v", "dep:v", "dep:versions"];
-  static override description =
-    "Get versions of all currently used dependencies";
+  static override description = "Get versions of all dependencies";
   static override examples = ["<%= config.bin %> <%= command.id %>"];
   static override flags = {
     ...baseFlags,
+    default: Flags.boolean({
+      description:
+        "Display default npm and cargo dependencies and their versions for current CLI version. Default npm dependencies are always available to be imported in Aqua",
+    }),
   };
   async run(): Promise<void> {
-    const { maybeFluenceConfig } = await initCli(
+    const { flags, maybeFluenceConfig } = await initCli(
       this,
       await this.parse(Versions)
     );
+
+    if (flags.default) {
+      return commandObj.log(
+        yamlDiffPatch(
+          "",
+          {},
+          {
+            "cli version": commandObj.config.version,
+            "default npm dependencies. Can be overridden with 'fluence dependency npm install <name>@<version>'":
+              getRecommendedDependencies("npm"),
+            "default cargo dependencies. Can be overridden with 'fluence dependency cargo install <name>@<version>'":
+              getRecommendedDependencies("cargo"),
+          }
+        )
+      );
+    }
 
     commandObj.log(
       yamlDiffPatch(
@@ -45,7 +68,7 @@ export default class Versions extends BaseCommand<typeof Versions> {
         {},
         {
           [`${CLI_NAME} version`]: commandObj.config.version,
-          "rust-peer version": versions["rust-peer"],
+          "nox version": versions["nox"],
           "rust toolchain": versions["rust-toolchain"],
           [`npm dependencies that can be overridden with \`${CLI_NAME} dependency npm install <name>@<version>\``]:
             await resolveDependencies("npm", maybeFluenceConfig),